Add files via upload
Hello, I require to sort through a table of information based on date and time, while taking into consideration other column variables. I would also like an explanation on how to do it as I am trying to learn this myself too. Need this done in python, using pandas.

The current scenario that I have is the table below. I have a data set for numerous days, and I would like to be see only the rows in which the high value at 10am is greater than the open value at 930am by at least 2 (ie, went from 16 to 18) and I would like to see in how many days this was true